fix some wars, fix mem leak, add unit test

+++ port/linux/package/pikascript/pikascript-lib/re/pcre_valid_utf8.c
@@ -1,4 +1,4 @@
-
+#include <stdint.h>
 #include "re_config.h"
 #include "pcre_internal.h"
 
@@ -37,7 +37,7 @@ register const uschar *p;
 if (length < 0)
   {
   for (p = string; *p != 0; p++);
-  length = p - string;
+  length = (uintptr_t)p - (uintptr_t)string;
   }
 
 for (p = string; length-- > 0; p++)
@@ -45,13 +45,13 @@ for (p = string; length-- > 0; p++)
   register int ab;
   register int c = *p;
   if (c < 128) continue;
-  if (c < 0xc0) return p - string;
+  if (c < 0xc0) return (uintptr_t)p - (uintptr_t)string;
   ab = _pcre_utf8_table4[c & 0x3f];     /* Number of additional bytes */
-  if (length < ab || ab > 3) return p - string;
+  if (length < ab || ab > 3) return (uintptr_t)p - (uintptr_t)string;
   length -= ab;
 
   /* Check top bits in the second byte */
-  if ((*(++p) & 0xc0) != 0x80) return p - string;
+  if ((*(++p) & 0xc0) != 0x80) return (uintptr_t)p - (uintptr_t)string;
 
   /* Check for overlong sequences for each different length, and for the
   excluded range 0xd000 to 0xdfff.  */
@@ -61,7 +61,7 @@ for (p = string; length-- > 0; p++)
     /* Check for xx00 000x (overlong sequence) */
 
     case 1:
-    if ((c & 0x3e) == 0) return p - string;
+    if ((c & 0x3e) == 0) return (uintptr_t)p - (uintptr_t)string;
     continue;   /* We know there aren't any more bytes to check */
 
     /* Check for 1110 0000, xx0x xxxx (overlong sequence) or
@@ -70,7 +70,7 @@ for (p = string; length-- > 0; p++)
     case 2:
     if ((c == 0xe0 && (*p & 0x20) == 0) ||
         (c == 0xed && *p >= 0xa0))
-      return p - string;
+      return (uintptr_t)p - (uintptr_t)string;
     break;
 
     /* Check for 1111 0000, xx00 xxxx (overlong sequence) or
@@ -80,7 +80,7 @@ for (p = string; length-- > 0; p++)
     if ((c == 0xf0 && (*p & 0x30) == 0) ||
         (c > 0xf4 ) ||
         (c == 0xf4 && *p > 0x8f))
-      return p - string;
+      return (uintptr_t)p - (uintptr_t)string;
     break;
 
 #if 0
@@ -90,13 +90,13 @@ for (p = string; length-- > 0; p++)
 
     /* Check for 1111 1000, xx00 0xxx */
     case 4:
-    if (c == 0xf8 && (*p & 0x38) == 0) return p - string;
+    if (c == 0xf8 && (*p & 0x38) == 0) return (uintptr_t)p - (uintptr_t)string;
     break;
 
     /* Check for leading 0xfe or 0xff, and then for 1111 1100, xx00 00xx */
     case 5:
     if (c == 0xfe || c == 0xff ||
-       (c == 0xfc && (*p & 0x3c) == 0)) return p - string;
+       (c == 0xfc && (*p & 0x3c) == 0)) return (uintptr_t)p - (uintptr_t)string;
     break;
 #endif
 
@@ -105,7 +105,7 @@ for (p = string; length-- > 0; p++)
   /* Check for valid bytes after the 2nd, if any; all must start 10 */
   while (--ab > 0)
     {
-    if ((*(++p) & 0xc0) != 0x80) return p - string;
+    if ((*(++p) & 0xc0) != 0x80) return (uintptr_t)p - (uintptr_t)string;
     }
   }
 #endif
